+++ tests/FastForwarderTest.php
@@ -54,8 +54,9 @@ class FastForwarderTest extends \PHPUnit_Framework_TestCase
             ->skipWhenNthDayOfWeekOfMonth(3, 1, 2, 'presidents_day')
             ->skipWhenNthDayOfWeekOfMonth(4, 4, 11, 'thanksgiving')
             ->skipWhenMonthAndDay(1, 1) // test auto-naming
+            ->skipWhen(['iansltx\BusinessDays\StaticFilter', 'isEasterMonday'], 'easter_monday')
             ->skipWhen(function (\DateTimeInterface $dt) { // test providing a callable directly
-            return $dt->format('m') == 12 && $dt->format('d') == 25;
-        }, 'christmas');
+                    return $dt->format('m') == 12 && $dt->format('d') == 25;
+                }, 'christmas');
     }
 }